feat: implement About and Login page components with client layout and add metadata.
