+++ demo_frontend/src/components/layout/ClientLayout.tsx
@@ -1,0 +1,81 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { usePathname, useRouter } from "next/navigation";
+
+import Header from "@/components/general/Header";
+import Footer from "@/components/general/Footer";
+import CursorTrail from "@/components/effects/CursorTrail";
+import { ThemeProvider } from "@/contexts/ThemeContext";
+import ConfirmationModal from "@/components/ui/ConfirmationModal";
+import Loader from "@/components/ui/Loader";
+import { logout } from "@/lib/api/auth";
+
+export default function ClientLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const router = useRouter();
+  const pathname = usePathname();
+
+  const [showConfirmation, setShowConfirmation] = useState(false);
+  const [isLoggingOut, setIsLoggingOut] = useState(false);
+
+  const handleLogoutClick = () => {
+    router.prefetch("/login");
+    setShowConfirmation(true);
+  };
+
+  const handleCancelLogout = () => setShowConfirmation(false);
+
+  const handleConfirmLogout = async () => {
+    setIsLoggingOut(true);
+    try {
+      await logout();
+      router.replace("/login");
+    } catch {
+      setIsLoggingOut(false);
+      setShowConfirmation(false);
+    }
+  };
+
+  useEffect(() => {
+    if (pathname === "/login" && (isLoggingOut || showConfirmation)) {
+      setIsLoggingOut(false);
+      setShowConfirmation(false);
+    }
+  }, [pathname, isLoggingOut, showConfirmation]);
+
+  return (
+    <ThemeProvider>
+      <CursorTrail />
+      <Header onLogoutClick={handleLogoutClick} isLoggingOut={isLoggingOut} />
+
+      <main className="flex-1 w-full flex flex-col justify-center items-center">
+        <div className="relative w-full h-full flex items-center justify-center">
+          {showConfirmation ? (
+            <ConfirmationModal
+              isOpen={showConfirmation}
+              title="Confirm Logout"
+              message="Are you sure you want to logout? You will need to login again to access your account."
+              confirmText={isLoggingOut ? "Logging out..." : "Yes, Logout"}
+              cancelText="Cancel"
+              onConfirm={handleConfirmLogout}
+              onCancel={handleCancelLogout}
+              confirmButtonClass="bg-red-500 hover:bg-red-600 text-white"
+              confirmDisabled={isLoggingOut}
+              cancelDisabled={isLoggingOut}
+            />
+          ) : (
+            children
+          )}
+
+          {isLoggingOut && <Loader cover="content" />}
+        </div>
+      </main>
+
+      <Footer />
+    </ThemeProvider>
+  );
+}
